requirements.txt
# Audio-to-RAG POC Dependencies
# Core AI/ML libraries
voyageai>=0.2.0
assemblyai>=0.23.0
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0

# Database
pymongo>=4.6.0

# Environment management
python-dotenv>=1.0.0

# Audio processing and recording
pyaudio>=0.2.11
wave

# Text-to-speech
pyttsx3>=2.90

# HTTP requests
requests>=2.28.0

# Python standard library enhancements (if needed)
# These are typically included with Python but listed for completeness
typing-extensions>=4.5.0
